Avoid negative sleep in pre-stop hook and exit on missing credentials (#6602)
GNU sleep does not allow negative numbers. I did a lot of testing for the original commit that introduced the new script on a Mac which is of BSD heritage and does not have a problem with that 😞

The second change in this PR exits the script if we don't have API credentials. This is the case during the migration to the script after the operator upgrade which will introduce the new logic: the config map with the script will auto-update after some time but the Pod definition will of course stay the same leading to a situation where we have the new script but don't have the credentials available yet because the necessary volume mounts are missing. We don't support running with xpack.security.enabled: false so I think this is an OK decision to make.