Logstash pipelines (#6480)
This PR allows configuration of Logstash pipeline.yml in CRD. pipeline.yml is an array of map of string/interface{} defining multiple pipelines. `.` in between the key is treated as string. Pipelines can be set from either inline in Pipelines or secret referencing in PipelinesRef.

e2e test: TestPipelineConfigRefLogstash, TestPipelineConfigLogstash

Co-authored-by: Rob Bavey <rob.bavey@elastic.co>
Co-authored-by: Michael Morello <michael.morello@gmail.com>