E2E: retry setup of data integrity check (#6710)
We have seen some test failures on builds where the cluster rejects connection attempts, even though it has previously already responded to health checks. The setup of the data integrity check is one of the few places where we do not retry requests on failures. I looked at the logs and it seems the requests are made only a few seconds after the cluster reached ready state so I am wondering if we should be more lenient here and allow for retries. 

Also track the internal http service in e2e test checks and fix a potential resource leak in the test client.