Fix the sample tests for logstash service and stack monitoring
The stack monitoring tests were failing as the builder was not creating the monitoring
elasticsearch references. This was added, as well as a minor refactoring to split the builder
into adding metrics and logging monitoring separately.

This commit also calculates the the monitoring assocation counts correctly, as previously
the metrics association was double counted.

This commit also fixes the logstash service sample end to end test - the pipeline id in the
yaml definition did not match that in the test.