Update Logstash stack mon tests (#6741)
This adjusts the number of volumes expected from Beats sidecars in the Logstash
Stack Monitoring unit tests.

Why? Because we don't test PRs with an automatic merge of the main branch (🐛🐞), we missed that the tests in #6732 had to be updated to take into account the changes made by #6703, which adds a new temp volume to the Beats sidecars.