Fix incorrect auth secret_token key post 8.0 (#6769)
Also changes the e2e tests to check the case where a client tries to use an incorrect auth token to make a requests (which should be rejected). This acts as a test for correct configuration of the APM server with a auth token. Because we did not do this check we missed the misconfiguration since 8.0 in our e2e testing.

Marking this as breaking because we shipped versions of ECK that misconfigured 8.0+ APM servers and therefore there might be user installations out there that work only because of this misconfiguration. If we now ship the fix and their clients are not actually setting an auth token then APM server will start rejecting their requests.