Allow to publish non SNAPSHOT Helm chart in dev Helm repo (#6792)
This adds another condition before setting the precondition that prevents to overwrite a Helm chart package when uploading to the GCS bucket that is behind a Helm repo. In addition to not being a snapshot, the release must now be done on the prod repo for the precondition to be added.