[buildkite] Limit notified failures to not exceed message longer than 3000 bytes (#6822)
This prevents to send a Slack notification with a message greater than 3000 chars by limiting the number of notified test failures to 15 and adding the text Go to build to see all xyz failures. at the end when there are more.

To make this workable from the template, the map of list of short failures by env is flattened and the total count of short failures is injected in the template.

Not related to this, this also renames trimError in truncateError for consistency.