Do not close channels, remove errexit from run.sh (#6918)
Avoid having to wait for log completion and avoid running into "send on closed channel" errors by not closing the error and log channel in the job abstraction we use for e2e tests. The rationale is that the program will exit anyway at this point and we do not need to manager those resources.

Remove the errexit shell option from run.sh to avoid the script exiting early on errors and preventing the job manager to download the log file.