Release parent charts with lock file (#6911) (#6932)
This restores the Chart.lock file in the published parent charts eck-operator and eck-stack
by doing the equivalent of helm update dependency, which is now safe as all dependencies
are local without repository. No index is updated locally.

This also adds a go-run make target and a flag --keep-tmp-dir to keep the temporary directory
and print its name to make this more convenient to develop.