Lazy vault client creation when reading vault file (#6927)
This allows reading a Vault file without creating a Vault client when the file already exists, which was the main purpose of the `vault.ReadFile` (reads the file if it exists or else reads the Secret from Vault and writes it to on disk).

A new `ClientProvider` type is introduced with an associated constructor and the signature of the `vault.ReadFile` func is updated to use it.

Motivation for this change: make the operator build (with the go generate step) possible in Docker, which opens the door to use the elastic docker-builder buildkite agent. By fetching the public key first, we can go generate during the Docker build (in the Dockerfile) without the dependency to Vault, which is a pain to manage in a Dockerfile.