Generate the trigger step to remove v prefix from BUILDKITE_TAG (#6964)
This fixes a bug introduced in #6879 where we can't just trigger the pipeline that copies the images with
IMAGES_TAG: "${BUILDKITE_TAG}" because the tag contains the v prefix and ECK images are tagged
without it. The fix is to go through a bit of bash to generate the step with the variable set without the v using shell parameter expansion.