Fix tests run dependency (#6970)
I got confused at the last moment in 445aee2 with the dependencies of the step to run e2e tests. We need to depend on the step that creates the cluster, the dependencies on the build steps were already removed.

It was not catch in the PR because we use kind for PRs which is a provider that doesn't use a remoteConfig.