E2E set test storage class for Logstash data volume (#6994)
Always set the storage class for Logstash to e2e-default in the e2e tests. This is a pragmatic deviation from the practice we have for Elasticsearch were we test with the e2e-default storage class in general but test with network attached volumes in samples and recipes to have both local and networked volume test paths.

Adding something similar for Logstash seemed to complex with the current builders we have.

Without this change Logstash tests fail on EKS where the EBS CSI driver is not installed by default.