use core-ubuntu-2004 image for gcp steps
The elastic-buildkite-agent-ubuntu-2004-lts image was one of the very
first hand-rolled images that we created when experimenting with
Buildkite and hasn't been updated/improved in quite some time. One of
its shortcomings is that it still has a filebeat/metricbeat
configuration that sends PagerDuty alerts when the VM's root volume goes
over 90% full, which causes unnecessary toil for our SRE colleagues.

Switch over to the newer, actively maintained core-ubuntu-2004 image,
which is built automatically every day and should provide the same
functionality as the previous image but without the PagerDuty alerts.