Do not fail build on k8s cluster deletion failure (#7197)
This updates the pipeline template used to create the pipeline that runs the e2e tests, to not fail the build when deleting a k8s cluster fails, since we now have a dedicated `pipeline cloud-on-k8s-operator-e2e-clusters-cleanup` which periodically retries these failed deletions. This also removes the retry mechanism that was here because it is completely inefficient, because lacking an exponential backoff delay.