Update PVC webhook check to allow storage request changes
Updates the PVCChanges webhook check to check whether a storage request
change has been made to a volumeClaimTemplate in the LogstashSpec.

This update does not do a full verifiation to check whether the
storage update is valid in the environment - this will be added
in a follow-up PR, as the check to compare storage requires the
use of the k8s client, which will require a rework of validation
to avoid an import cycle.