Use readiness port (#7847)
As of 8.2.0 use the new readiness.port setting to enable a TCP readiness check that is sensitive to cluster membeship of the node. This should improve cluster availablity during external disruptions e.g. due to node upgrades, as the PDB will disallow furhter upgrades until the most recently upgraded node has rejoined the cluster.

I choose to create a separate file mounted next to the existing script to facilitate upgrades. Specifically I wanted to avoid a situation where during the upgrade we update the scripts configmap overwriting the old probe script with the new one and thereby breaking the not yet upgraded nodes in the cluster.

The approach taken here creates some techical debt in form of the extra script hanging around when not needed. But I imaging we can drop it when we stop supporting ES < 8.2.0.

Marked as bug because a cluster might be unavailable if master nodes are deleted while the previously deleted ones are not yet back in the cluster, which violates our promise of interruption free ES operations when running in HA mode with multiple master nodes.

Because there is an edge case with single node clusters or clusters that have all node unavailable at the same time a new URL provider concept has been introduced that targets running pods (not necessarily ready ones even though these are preferred) when making calls to ES APIs. 

The edge case addressed is: 
* the readiness.port reports an ES node as not ready if a shutdown record exists for the ES node
* an unready node is not listed in the (internal) k8s service we use and assumed not reacheable
* this means in turn that the operator is not able to call the DELETE _nodes/$NODE_ID/shutdown API to remove the shutdown record and the cluster remains unready forever

By not going through the k8s service abstraction we can target individual running Pods that are not ready yet and delete shutdown records to reach a ready state.