Fix panic and dropped error result (#7875)
We were ignoring an error during a node shutdown initialisation and panicking on a subsequently uninitialised map. 

Technically it should be enough to handle the error but I chose to also intialise the map by default. 

For future reference, this was discovered through an e2e test (props to @thbkrkr for the throrough analysis). 
```
[
	expected less than 1, got 2, restarts: [
		{o44cyWaAQQuzmvh3g_iOzQ RESTART 187844 1717724317347 COMPLETE {COMPLETE 0 no shard relocation is necessary for a node restart} {COMPLETE} {COMPLETE}} 
		{LSN_PaSuRoOIsQF3QKyx3g RESTART 187995 1717724469546 COMPLETE {COMPLETE 0 no shard relocation is necessary for a node restart} {COMPLETE} {COMPLETE}}
]
```
We discovered that a previous shutdown record was still present. This should not happen as we clear shutdown records before proceeding to the next node. We then found that the operator had been restarted twice once in the middle to the upgrade process. Which lead to the discovery of a panic in the node.go when writing to a nil map.