ECK S3 wrong truststore name in example
Hello team!

I received customer feedback about this, and among other things that I'm adding as an issue to the documentation team, they mentioned this. 

In step 4, the secret name is created as `custom-truststore` but in the sample manifest, it says `additional-certs`.

I hope this fix can be merged soon! Thanks!