Update stack monitoring documentation page (#8198)
This updates the stack monitoring documentation page.

It having 5 notes in a row, makes it hard to prioritize them. So I removed them and replace them with text.

> You can configure an Elasticsearch cluster to monitor itself.

Moved to the start of the page, including an advice to avoid it

> If Stack Monitoring is configured for a Beat, but the corresponding Elasticsearch cluster is not monitored, the Kibana Stack Monitoring page will not show the Beats data.

Moved to the end of the page where similar is stated for Kibana

> Stack Monitoring for Logstash on ECK is only available for Logstash versions 8.7.0 and later.

Removed this statement from this page. It is not just the stack monitoring that is not supported, it is the using Logstash at all in ECK lower than 8.7.0.

> The monitoring cluster must be managed by ECK in the same Kubernetes cluster as the monitored one.

This statement is conflicting with the next doc page that is "Connect to an external monitoring Elasticsearch cluster". It is true for the example on the page, so I added an introductory line for the example

---------
Co-authored-by: Thibault Richard <thbkrkr@users.noreply.github.com>