Logstash e2e: ssl => ssl_enabled and fix supported versions listing (#8460)
Between 8.7 and 8.8 ssl has been renamed to ssl_enabled as of 9.0.0 ssl is not longer allowed. This removes the attribute from the test config as it is implied by the https scheme in the hosts URL.

This updates the property to ssl_enabled. Also we do incorrectly state that we support Logstash 8.7+ while in code we exclude anything before 8.12!