Re-enable 4.12 while Red Hat provides long term support (#8552)
Based on https://access.redhat.com/support/policy/updates/openshift#ocp4 4.12 will receive security updates until 2026. While we do not support it anymore officially this changes the OLM config to avoid locking out users of this version from upgrading to 3.0.

This is to break up a dependency chain where an upgarde to stack 9.0 would be predicated on an upgrade to ECK 3.0 which for users on 4.12 would first require a Kubernetes upgrade. While API compatibility is a concern as 1.25 is officially unsupported from a Kubernetes point of view there are currently no known technical blockers that prevent ECK 3.0 from running on 1.25 at the users discretion.

To be clear this does not change our official stance regarding the EOL state of OCP 4.12 it just removes the technical hurdle that would prevent users from running ECK 3.0 on that version.