Update docs-build.yml to build version branches too (#8668) (#8670)
This should be back-ported to `3.0` if the intend is to publish the `3.0` branch to elastic.co/docs instead of `main`.

After this is backported and our workflows ran on the respective branches we can switch the documentation configuration over to use `3.0` as the source for production builds.

(cherry picked from commit 4f714469aed46faf5fb0d2331384cc79ccf85246)

Co-authored-by: Martijn Laarman <Mpdreamz@gmail.com>