Move from deprecated container input to filestream (#8679)
Replace the deprecated container input with a filestream input on the ECK operator StatefulSet/Deployment in both OLM and standalone ECK.

I also noticed that the ~ YAML notation for null did break the configuration when using kubectl apply for some reason that have not dug deep enough to understand. Moving to {} instead led to a working configuration. I therefore also applied this change to the relevant configuration recipes.