Remove one master at a time upscale restriction (#8940)
#8939 describes how requests to Elasticsearch are made during upscales. When scaling an overloaded ES cluster this can be problematic as the cluster won't be able to serve these requests in time.

This change removes the source of those requests: the on master at a time scale up restriction. This functionality stems for a time when the ECK operator had to manually reconfigure the quorum during scale up (Elasticsearch 6 and earlier). We no longer support these early versions where a quick scale up could lead to potential split brain scenarios if a partition was formed during scale up.

By removing the one-at-at-time invariant the call to the _nodes API during scale up becomes unnecessary. The call is now only made when calculating the scale downs.