Helm chart: clarify elasticsearchRef usage in Kibana http-configuration.yaml example (#8959)
This example file is referenced in this doc, https://www.elastic.co/docs/deploy-manage/deploy/cloud-on-k8s/managing-deployments-using-helm-chart#k8s-eck-stack-helm-customize, but the `elasticsearchRef` is incorrect.

It's incorrect because the Elasticsearch resource name in these cases is hardcoded by the chart to `elasticsearch` and it's not using the full name of the installed release (which would be ideal).

Additionally, IMO the user should not need to touch the reference when deploying Elasticsearch and Kibana together with the `eck-stack` helm chart, and should allow the chart to configure the reference. So the best solution here is to just leave it commented out (what I'm proposing), and not set it statically to any value.

Please correct me if I'm wrong, as I could be missing something.

I've commented out that reference and included some comments, as otherwise the example from [the official doc](https://www.elastic.co/docs/deploy-manage/deploy/cloud-on-k8s/managing-deployments-using-helm-chart#k8s-eck-stack-helm-customize) fails.

If this example is used in other places where the static reference is mandatory we need to solve this in a different way. The fact is that our published documented procedure fails because of this.